Moved CommandLineParser tests to the same package
